Fixed classes Laps Tracks and Attributes AHR in activity

+++ FitnessManager/src/main/java/gr/hua/dit/fitnessmanager/Activity.java
@@ -25,12 +25,12 @@ public abstract class Activity {
         return getTimeSeconds() / 60 / (getDistanceMeters() / 1000);
     }
 
-    public double getAHR() {
-        int sum = 0;
+        public int getAHR() {
+            int sum = 0;
         for (Laps lap: laps) {
             sum += lap.getAHR();
         }
-        return (double) sum / laps.size();
+        return sum / laps.size();
     }
 
     public double getAverageSpeed() {
